re-add code to delete test schema in class teardown and use integer field for some stat fields BB-2676

+++ bluebottle/statistics/serializers.py
@@ -6,10 +6,10 @@ from .models import Statistic
 
 class StatisticSerializer(serializers.ModelSerializer):
     donated = serializers.DecimalField(source='donated')
-    projects_online = serializers.DecimalField(source='projects_online')
-    projects_realized = serializers.DecimalField(source='projects_realized')
-    tasks_realized = serializers.DecimalField(source='tasks_realized')
-    people_involved = serializers.DecimalField(source='people_involved')
+    projects_online = serializers.IntegerField(source='projects_online')
+    projects_realized = serializers.IntegerField(source='projects_realized')
+    tasks_realized = serializers.IntegerField(source='tasks_realized')
+    people_involved = serializers.IntegerField(source='people_involved')
 
     class Meta:
         model = Statistic
